+++ DotDoc.EntityFrameworkCore.Extensions/Extensions/GetContextExtensions.cs
@@ -8,7 +8,6 @@ using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Query;
 using Microsoft.EntityFrameworkCore.Query.Internal;
 using Microsoft.EntityFrameworkCore.Storage;
-using System.Data;
 using System.Diagnostics.CodeAnalysis;
 using System.Reflection;