+++ DotDoc.EntityFrameworkCore.Extensions/Extensions/QueryExtensions.cs
@@ -198,10 +198,10 @@ public static class QueryExtensions
     /// <param name="sql">The <see cref="FormattableString"/> representing a SQL query with parameters.</param>
     /// <param name="page">Page number to return (starting at 0).</param>
     /// <param name="pageSize">Number of records per page.</param>
-    /// <returns>An instance of <see cref="QueryPage"/> containing the page data (If the page number is past the end of the table then the it will become the last page).</returns>
-    public static QueryPage ExecutePagedQuery(this DatabaseFacade databaseFacade, FormattableString sql, long page, long pageSize)
+    /// <returns>An instance of <see cref="QueryPage{DataTable}"/> containing the page data (If the page number is past the end of the table then the it will become the last page).</returns>
+    public static QueryPage<DataTable> ExecutePagedQuery(this DatabaseFacade databaseFacade, FormattableString sql, long page, long pageSize)
     {
-        QueryPage queryPage = QueryMethods.ExecutePagedQuery(databaseFacade, sql.Format, sql.GetArguments(), page, pageSize);
+        QueryPage<DataTable> queryPage = QueryMethods.ExecutePagedQuery(databaseFacade, sql.Format, sql.GetArguments(), page, pageSize);
         return queryPage;
     }
 
@@ -213,10 +213,43 @@ public static class QueryExtensions
     /// <param name="page">Page number to return (starting at 0).</param>
     /// <param name="pageSize">Number of records per page.</param>
     /// <param name="parameters">Parameters to use with the SQL.</param>
-    /// <returns>An instance of <see cref="QueryPage"/> containing the page data (If the page number is past the end of the table then the it will become the last page).</returns>
-    public static QueryPage ExecutePagedQuery(this DatabaseFacade databaseFacade, string sql, long page, long pageSize, params IEnumerable<object> parameters)
+    /// <returns>An instance of <see cref="QueryPage{DataTable}"/> containing the page data (If the page number is past the end of the table then the it will become the last page).</returns>
+    public static QueryPage<DataTable> ExecutePagedQuery(this DatabaseFacade databaseFacade, string sql, long page, long pageSize, params IEnumerable<object> parameters)
     {
-        QueryPage queryPage = QueryMethods.ExecutePagedQuery(databaseFacade, sql, parameters, page, pageSize);
+        QueryPage<DataTable> queryPage = QueryMethods.ExecutePagedQuery(databaseFacade, sql, parameters, page, pageSize);
+        return queryPage;
+    }
+
+    /// <summary>
+    /// Executes a query and returns the specified page of results.
+    /// </summary>
+    /// <typeparam name="TEntity">Type of entity to return.</typeparam>
+    /// <param name="databaseFacade">The <see cref="DatabaseFacade"/> for the context.</param>
+    /// <param name="sql">The <see cref="FormattableString"/> representing a SQL query with parameters.</param>
+    /// <param name="page">Page number to return (starting at 0).</param>
+    /// <param name="pageSize">Number of records per page.</param>
+    /// <returns>An instance of <see cref="QueryPage{T}"/> where T is <see cref="IList{TEntity}"/> containing the page data (If the page number is past the end of the table then the it will become the last page).</returns>
+    public static QueryPage<IList<TEntity>> ExecutePagedQuery<TEntity>(this DatabaseFacade databaseFacade, FormattableString sql, long page, long pageSize)
+        where TEntity : class
+    {
+        QueryPage<IList<TEntity>> queryPage = QueryMethods.ExecutePagedQuery<TEntity>(databaseFacade, sql.Format, sql.GetArguments(), page, pageSize);
+        return queryPage;
+    }
+
+    /// <summary>
+    /// Executes a query and returns the specified page of results.
+    /// </summary>
+    /// <typeparam name="TEntity">Type of entity to return.</typeparam>
+    /// <param name="databaseFacade">The <see cref="DatabaseFacade"/> for the context.</param>
+    /// <param name="sql">The SQL query to execute.</param>
+    /// <param name="page">Page number to return (starting at 0).</param>
+    /// <param name="pageSize">Number of records per page.</param>
+    /// <param name="parameters">Parameters to use with the SQL.</param>
+    /// <returns>An instance of <see cref="QueryPage{T}"/> where T is <see cref="IList{TEntity}"/> containing the page data (If the page number is past the end of the table then the it will become the last page).</returns>
+    public static QueryPage<IList<TEntity>> ExecutePagedQuery<TEntity>(this DatabaseFacade databaseFacade, string sql, long page, long pageSize, params IEnumerable<object> parameters)
+        where TEntity : class
+    {
+        QueryPage<IList<TEntity>> queryPage = QueryMethods.ExecutePagedQuery<TEntity>(databaseFacade, sql, parameters, page, pageSize);
         return queryPage;
     }
 
@@ -228,10 +261,10 @@ public static class QueryExtensions
     /// <param name="page">Page number to return (starting at 0).</param>
     /// <param name="pageSize">Number of records per page.</param>
     /// <param name="cancellationToken">A <see cref="CancellationToken"/> to observe while waiting for the task to complete.</param>
-    /// <returns>An instance of <see cref="QueryPage"/> containing the page data (If the page number is past the end of the table then the it will become the last page).</returns>
-    public static async Task<QueryPage> ExecutePagedQueryAsync(this DatabaseFacade databaseFacade, FormattableString sql, long page, long pageSize, CancellationToken cancellationToken = default)
+    /// <returns>An instance of <see cref="QueryPage{DataTable}"/> containing the page data (If the page number is past the end of the table then the it will become the last page).</returns>
+    public static async Task<QueryPage<DataTable>> ExecutePagedQueryAsync(this DatabaseFacade databaseFacade, FormattableString sql, long page, long pageSize, CancellationToken cancellationToken = default)
     {
-        QueryPage queryPage = await QueryMethods.ExecutePagedQueryAsync(databaseFacade, sql.Format, sql.GetArguments(), page, pageSize, cancellationToken).ConfigureAwait(false);
+        QueryPage<DataTable> queryPage = await QueryMethods.ExecutePagedQueryAsync(databaseFacade, sql.Format, sql.GetArguments(), page, pageSize, cancellationToken).ConfigureAwait(false);
         return queryPage;
     }
 
@@ -244,10 +277,45 @@ public static class QueryExtensions
     /// <param name="pageSize">Number of records per page.</param>
     /// <param name="cancellationToken">A <see cref="CancellationToken"/> to observe while waiting for the task to complete.</param>
     /// <param name="parameters">Parameters to use with the SQL.</param>
-    /// <returns>An instance of <see cref="QueryPage"/> containing the page data (If the page number is past the end of the table then the it will become the last page).</returns>
-    public static async Task<QueryPage> ExecutePagedQueryAsync(this DatabaseFacade databaseFacade, string sql, long page, long pageSize, CancellationToken cancellationToken = default, params IEnumerable<object> parameters)
+    /// <returns>An instance of <see cref="QueryPage{T}"/> where T is <see cref="IList{TEntity}"/> containing the page data (If the page number is past the end of the table then the it will become the last page).</returns>
+    public static async Task<QueryPage<DataTable>> ExecutePagedQueryAsync(this DatabaseFacade databaseFacade, string sql, long page, long pageSize, CancellationToken cancellationToken = default, params IEnumerable<object> parameters)
     {
-        QueryPage queryPage = await QueryMethods.ExecutePagedQueryAsync(databaseFacade, sql, parameters, page, pageSize, cancellationToken).ConfigureAwait(false);
+        QueryPage<DataTable> queryPage = await QueryMethods.ExecutePagedQueryAsync(databaseFacade, sql, parameters, page, pageSize, cancellationToken).ConfigureAwait(false);
+        return queryPage;
+    }
+
+    /// <summary>
+    /// Executes a query and returns the specified page of results.
+    /// </summary>
+    /// <typeparam name="TEntity">Type of entity to return.</typeparam>
+    /// <param name="databaseFacade">The <see cref="DatabaseFacade"/> for the context.</param>
+    /// <param name="sql">The <see cref="FormattableString"/> representing a SQL query with parameters.</param>
+    /// <param name="page">Page number to return (starting at 0).</param>
+    /// <param name="pageSize">Number of records per page.</param>
+    /// <param name="cancellationToken">A <see cref="CancellationToken"/> to observe while waiting for the task to complete.</param>
+    /// <returns>An instance of <see cref="QueryPage{T}"/> where T is <see cref="IList{TEntity}"/> containing the page data (If the page number is past the end of the table then the it will become the last page).</returns>
+    public static async Task<QueryPage<IList<TEntity>>> ExecutePagedQueryAsync<TEntity>(this DatabaseFacade databaseFacade, FormattableString sql, long page, long pageSize, CancellationToken cancellationToken = default)
+        where TEntity : class
+    {
+        QueryPage<IList<TEntity>> queryPage = await QueryMethods.ExecutePagedQueryAsync<TEntity>(databaseFacade, sql.Format, sql.GetArguments(), page, pageSize, cancellationToken).ConfigureAwait(false);
+        return queryPage;
+    }
+
+    /// <summary>
+    /// Executes a query and returns the specified page of results.
+    /// </summary>
+    /// <typeparam name="TEntity">Type of entity to return.</typeparam>
+    /// <param name="databaseFacade">The <see cref="DatabaseFacade"/> for the context.</param>
+    /// <param name="sql">The SQL query to execute.</param>
+    /// <param name="page">Page number to return (starting at 0).</param>
+    /// <param name="pageSize">Number of records per page.</param>
+    /// <param name="cancellationToken">A <see cref="CancellationToken"/> to observe while waiting for the task to complete.</param>
+    /// <param name="parameters">Parameters to use with the SQL.</param>
+    /// <returns>An instance of <see cref="QueryPage{DataTable}"/> containing the page data (If the page number is past the end of the table then the it will become the last page).</returns>
+    public static async Task<QueryPage<IList<TEntity>>> ExecutePagedQueryAsync<TEntity>(this DatabaseFacade databaseFacade, string sql, long page, long pageSize, CancellationToken cancellationToken = default, params IEnumerable<object> parameters)
+        where TEntity : class
+    {
+        QueryPage<IList<TEntity>> queryPage = await QueryMethods.ExecutePagedQueryAsync<TEntity>(databaseFacade, sql, parameters, page, pageSize, cancellationToken).ConfigureAwait(false);
         return queryPage;
     }