+++ DotDoc.EntityFrameworkCore.Extensions/Model/QueryPage.cs
@@ -2,30 +2,29 @@
 // This file is licensed to you under the MIT license.
 // See the License.txt file in the solution root for more information.
 
-using System.Data;
-
 namespace DotDoc.EntityFrameworkCore.Extensions.Model;
 
 /// <summary>
 /// A page of query data. Returned by one of the ExecutePagedQueryxxxxx extensions.
 /// </summary>
-public class QueryPage
+/// <typeparam name="T">The type of page data.</typeparam>
+public class QueryPage<T>
 {
     /// <summary>
-    /// Initializes a new instance of the <see cref="QueryPage"/> class.
+    /// Initializes a new instance of the <see cref="QueryPage{T}"/> class.
     /// </summary>
     /// <param name="page">The page number.</param>
     /// <param name="pageSize">The page size.</param>
     /// <param name="recordCount">The number of records.</param>
     /// <param name="pageCount">The number of pages.</param>
-    /// <param name="dataTable">A <see cref="DataTable"/> containing the page.</param>
-    internal QueryPage(long page, long pageSize, long recordCount, long pageCount, DataTable dataTable)
+    /// <param name="pageData">The page data.</param>
+    internal QueryPage(long page, long pageSize, long recordCount, long pageCount, T pageData)
     {
         this.Page = page;
         this.PageSize = pageSize;
         this.RecordCount = recordCount;
         this.PageCount = pageCount;
-        this.DataTable = dataTable;
+        this.PageData = pageData;
     }
 
     #region public properties
@@ -51,9 +50,9 @@ public class QueryPage
     public long PageCount { get; }
 
     /// <summary>
-    /// Gets a <see cref="DataTable"/> containing the page.
+    /// Gets the page data.
     /// </summary>
-    public DataTable DataTable { get; }
+    public T PageData { get; }
 
     #endregion public properties
 }